requirements/development.txt
-r base.txt

ipdb==0.10.2
pdbpp==0.9.2
ipython==6.2.1

django-coverage-plugin==1.5.0
django-extensions==1.9.9
django-debug-toolbar==1.9.1

coverage==4.4.2
flake8==3.5.0
pytest==3.3.2
pytest-django==3.1.2
pytest-sugar==0.9.0
pytest-cov==2.5.1